requirements.txt
pyyaml
natsort
Pillow==10.0.1
exifread
peewee==3.17.0